# openapi.yaml
openapi: 3.0.0
info:
  title: MindPal API
  description: Complete documentation of all API endpoints for the MindPal Mental Health Application.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /profile:
    post:
      tags: [Profile]
      summary: Create profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                personalInfo:
                  type: object
                  properties:
                    firstName:
                      type: string
                      example: John
                    lastName:
                      type: string
                      example: Doe
                    dateOfBirth:
                      type: string
                      format: date
                      example: "1995-06-15"
                    gender:
                      type: string
                      enum: [male, female, other, prefer-not-to-say]
                      example: male
                    profilePicture:
                      type: string
                      example: "https://example.com/photo.jpg"
                contactInfo:
                  type: object
                  properties:
                    email:
                      type: string
                      example: john@example.com
                    phone:
                      type: string
                      example: "+1234567890"
                    address:
                      type: object
                      properties:
                        street:
                          type: string
                          example: "123 Main St"
                        city:
                          type: string
                          example: "New York"
                        state:
                          type: string
                          example: "NY"
                        country:
                          type: string
                          example: "US"
                        zipCode:
                          type: string
                          example: "10001"
                preferences:
                  type: object
                  properties:
                    theme:
                      type: string
                      enum: [light, dark, system]
                      example: system
                    language:
                      type: string
                      example: en
                    notificationPreferences:
                      type: object
                      properties:
                        email:
                          type: boolean
                          example: true
                        push:
                          type: boolean
                          example: true
                        sms:
                          type: boolean
                          example: false
                healthInfo:
                  type: object
                  properties:
                    height:
                      type: number
                      description: Height in cm
                      example: 175
                    weight:
                      type: number
                      description: Weight in kg
                      example: 70
                    bloodType:
                      type: string
                      enum: ["A+", "A-", "B+", "B-", "AB+", "AB-", "O+", "O-"]
                      example: "O+"
                    allergies:
                      type: array
                      items:
                        type: string
                      example: ["peanuts"]
                    medications:
                      type: array
                      items:
                        type: string
                      example: []
                    conditions:
                      type: array
                      items:
                        type: string
                      example: []
      responses:
        "201":
          description: Profile created

    get:
      tags: [Profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile details

    put:
      tags: [Profile]
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Same structure as create profile - all fields are optional for update
              properties:
                personalInfo:
                  type: object
                  properties:
                    firstName:
                      type: string
                      example: Jane
                    lastName:
                      type: string
                      example: Doe
                preferences:
                  type: object
                  properties:
                    theme:
                      type: string
                      enum: [light, dark, system]
                      example: dark
      responses:
        "200":
          description: Profile updated

    delete:
      tags: [Profile]
      summary: Delete profile
      responses:
        "200":
          description: Profile deleted
...
  /mood:
    post:
      tags: [Mood]
      summary: Record mood entry
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                moodValue:
                  type: integer
                  minimum: 1
                  maximum: 10
                  description: "1-10 scale where 10 is best mood"
                  example: 7
                moodLabel:
                  type: string
                  enum:
                    [
                      depressed,
                      sad,
                      neutral,
                      content,
                      happy,
                      excited,
                      anxious,
                      angry,
                      stressed,
                    ]
                  example: content
                factors:
                  type: object
                  properties:
                    sleepQuality:
                      type: integer
                      minimum: 1
                      maximum: 5
                      example: 4
                    stressLevel:
                      type: integer
                      minimum: 1
                      maximum: 5
                      example: 2
                    energyLevel:
                      type: integer
                      minimum: 1
                      maximum: 5
                      example: 4
                    socialInteraction:
                      type: integer
                      minimum: 1
                      maximum: 5
                      example: 3
                activities:
                  type: array
                  items:
                    type: string
                    enum:
                      [
                        meditation,
                        exercise,
                        work,
                        social,
                        creative,
                        rest,
                        journaling,
                        music,
                        nature,
                      ]
                  example: ["exercise", "social"]
                notes:
                  type: string
                  maxLength: 500
                  example: Feeling good after a morning run
      responses:
        "201":
          description: Mood recorded
...
  /journal:
    post:
      tags: [Journal]
      summary: Create journal entry
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Morning Reflection
                content:
                  type: string
                  example: Today I woke up feeling refreshed and ready for the day.
                moodBefore:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 3
                moodAfter:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 4
                tags:
                  type: array
                  items:
                    type: string
                    enum:
                      [
                        reflection,
                        gratitude,
                        stress,
                        anxiety,
                        happiness,
                        goal,
                        dream,
                      ]
                  example: ["reflection", "gratitude"]
                musicPlayed:
                  type: object
                  properties:
                    trackId:
                      type: string
                    name:
                      type: string
                      example: Calm Waters
                    artist:
                      type: string
                      example: Relaxation Music
                    duration:
                      type: number
                    mood:
                      type: string
                images:
                  type: array
                  items:
                    type: string
                audio:
                  type: string
                  format: binary
      responses:
        "201":
          description: Journal created
...
  /goals:
    get:
      tags: [Goals]
      summary: Get all goals
      responses:
        "200":
          description: List of goals
    post:
      tags: [Goals]
      summary: Create goal
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Exercise daily for 30 minutes
                description:
                  type: string
                  example: Build a consistent exercise habit
                category:
                  type: string
                  enum:
                    [
                      mental-health,
                      productivity,
                      relationships,
                      fitness,
                      learning,
                      other,
                    ]
                  example: fitness
                difficulty:
                  type: string
                  enum: [easy, medium, hard]
                  example: medium
                targetDate:
                  type: string
                  format: date
                  example: "2026-06-01"
                milestones:
                  type: array
                  items:
                    type: object
                    properties:
                      description:
                        type: string
                  example:
                    - description: "Complete first week"
                    - description: "Complete first month"
      responses:
        "201":
          description: Goal created
...
  /therapists:
    get:
      tags: [Therapists]
      summary: Search/list therapists
      security: []
      parameters:
        - in: query
          name: specialization
          schema:
            type: string
        - in: query
          name: location
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: List of therapists
...
  /resources:
    get:
      tags: [Resources]
      summary: Get resources
      parameters:
        - in: query
          name: type
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Resource list
...
  /search:
    get:
      tags: [Search]
      summary: Global search
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: type
          schema:
            type: string
      responses:
        "200":
          description: Search results
...